\name{awspack-package}
\alias{awspack-package}
\alias{awspack}
\docType{package}
\title{Amazon Web Services Bundle Package}
\description{A simple bundler of the cloudyr suite of package for Amazon Web Services (AWS). It imports all functions from all of the AWS packages into its own namespace and exports them. If you only want functions from a specific package, load the individual package instead.}
%\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
