\name{awspack-package}
\alias{awspack-package}
\alias{awspack}
\docType{package}
\title{Amazon Web Services Bundle Package}
\description{A simple bundler of \href{http://cloudyr.github.io}{the cloudyr project} suite of packages for Amazon Web Services (AWS). It imports all functions from all of the AWS packages into its own namespace and exports them. If you only want functions from a specific package, load the individual package instead.}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
\keyword{ package }
